% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_Model}
\alias{Rcpp_Model}
\alias{Model}
\alias{Rcpp_Model-class}
\title{Model Class}
\description{
Internal representation for multi-server job model with random speed scaling.
}
\section{Slots}{

\describe{
\item{\code{rho}}{Returns rho computated by using Neuts ergodicity criteria.}

\item{\code{mean_clients}}{Returns mean clients in system.}

\item{\code{mean_queue}}{Returns mean queue length.}

\item{\code{sum_pi_from_c_to_inf}}{Returns sum of distribution vectors from level N to infinity level.}

\item{\code{pi_0_c}}{Returns distribution for first N levels.}

\item{\code{distribution}}{\itemize{
\item Parameter: max_level - maximum level for which you want to calculate distribution;
\item Returns: list of vectors containing stationary distribution.
}}

\item{\code{level_description}}{\itemize{
\item Parameter: level -  level for which you want to get a description;
\item Returns: human-readable description of level in form "m|s_1,s_2,..." where "m" is a speed mode and "s_i" is
a number of serviced customers of "i" class.
}}

\item{\code{level_busy_servers}}{\itemize{
\item Parameter: level -  level for which you want to get a description;
\item Returns: vector of numbers of busy servers for each state at specified level.
}}

\item{\code{level_serviced_clients}}{\itemize{
\item Parameter: level -  level for which you want to get a description;
\item Returns: vector of numbers of serviced servers for each state at specified level.
}}

\item{\code{transient_analysis}}{\itemize{
\item Parameter: order - numerical order of Taylor series method
\item Parameter: stepsize - If positive, then  step length of the algorithm. If negative - the multiplier of the 
maximum step of the algorithm.
\item Returns: internal representation for model transient analysis (ModelTransient class).
}}
}}

