% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_Model$distribution}
\alias{Rcpp_Model$distribution}
\alias{Model$distribution}
\alias{Rcpp_Model-class$distribution}
\title{Calculate stationary distribution}
\arguments{
\item{max_level}{A maximum level for which you want to calculate distribution.}
}
\value{
A list of vectors containing stationary distribution.
}
\description{
Calculate stationary distribution
}
