% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_Model$level_busy_servers}
\alias{Rcpp_Model$level_busy_servers}
\alias{Model$level_busy_servers}
\alias{Rcpp_Model-class$level_busy_servers}
\title{Busy servers for level}
\arguments{
\item{level}{A level for which you want to get a description;}
}
\value{
A vector of numbers of busy servers for each state at specified level.
}
\description{
Busy servers for level
}
