% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_Model$level_description}
\alias{Rcpp_Model$level_description}
\alias{Model$level_description}
\alias{Rcpp_Model-class$level_description}
\title{Human-readeble level description}
\arguments{
\item{level}{A level for which you want to get a description;}
}
\value{
A human-readable description of level in form "m|s_1,s_2,..." where "m" is a speed mode and "s_i" is
a number of serviced customers of "i" class
}
\description{
Human-readeble level description
}
