% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_Model$level_serviced_clients}
\alias{Rcpp_Model$level_serviced_clients}
\alias{Model$level_serviced_clients}
\alias{Rcpp_Model-class$level_serviced_clients}
\title{Serviced clients}
\arguments{
\item{level}{A level for which you want to get a description;}
}
\value{
A vector of numbers of serviced clients for each state at specified level.
}
\description{
Serviced clients
}
