% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_Model$transient_analysis}
\alias{Rcpp_Model$transient_analysis}
\alias{Model$transient_analysis}
\alias{Rcpp_Model-class$transient_analysis}
\title{Transient analysis}
\arguments{
\item{order}{A numerical order of Taylor series method.}

\item{stepsize}{If positive, then  step length of the algorithm. If negative - the multiplier of the 
maximum step of the algorithm}
}
\value{
An internal representation for model transient analysis (ModelTransient class)
}
\description{
Transient analysis
}
