% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_ModelTransient$distribution}
\alias{Rcpp_ModelTransient$distribution}
\alias{ModelTransient$distribution}
\alias{Rcpp_ModelTransient-class$distribution}
\title{Transient distribution}
\arguments{
\item{max_time}{A maximum time for which you want to calculate mean customers in system;}

\item{pi_0}{A list of vectors contains distribution at zero time.}
}
\value{
A list of lists of vectors containing transient distribution at times proportional to the step length h.
}
\description{
Transient distribution
}
