% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_ModelTransient$mean_clients}
\alias{Rcpp_ModelTransient$mean_clients}
\alias{ModelTransient$mean_clients}
\alias{Rcpp_ModelTransient-class$mean_clients}
\title{Mean customers in system}
\arguments{
\item{max_time}{A maximum time for which you want to calculate mean customers in system;}

\item{pi_0}{A list of vectors contains distribution at zero time.}
}
\value{
A vector contains mean clients in system at times proportional to the step length h.
}
\description{
Mean customers in system
}
