% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_num_of_threads}
\alias{set_num_of_threads}
\title{set_num_of_threads}
\usage{
set_num_of_threads(threads)
}
\arguments{
\item{threads}{OpenMP only: number of threads for computation.}
}
\value{
TRUE, if OpenMP is enabled, FALSE otherwise.
}
\description{
Set number of threads. Works only if OpenMP is enabled. Max number of threads restricted by your CPU
}
