% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_model}
\alias{build_model}
\title{build_model}
\usage{
build_model(lambda, N, class_desc, f, P_a, P_d)
}
\arguments{
\item{lambda}{Input rate.}

\item{N}{Number of servers.}

\item{class_desc}{Description for clients class, i.e. matrix where first row is server count, second row is probability, last row is mu.}

\item{f}{Speed in each modes.}

\item{P_a}{Mode switching stochastic matrix when arrivals.}

\item{P_d}{Mode switching stochastic matrix when departures.}
}
\value{
Class that describes model.
}
\description{
Built multiserver job model with random speed scaling
}
