% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_model}
\alias{build_model}
\title{build_model}
\usage{
build_model(lambda, N, class_desc, f, P_a, P_d)
}
\arguments{
\item{lambda}{Input rate.}

\item{N}{Number of servers.}

\item{class_desc}{Description for clients class, i.e. matrix where first row is server count, second row is probability, last row is mu.}

\item{f}{Speed in each modes.}

\item{P_a}{Mode switching stochastic matrix when arrivals.}

\item{P_d}{Mode switching stochastic matrix when departures.}
}
\value{
Class that describes model.
}
\description{
Built multiserver job model with random speed scaling
}
\examples{
library(MJMrss)
lambda = 10 # Input rate
N = 10 # Number of servers
f = c(1, 2.2) # Speed (frequency)
P_a = matrix(c(0.1, 0.9,
               0.0, 1.0), nrow = 2, byrow = TRUE)
P_d = matrix(c(1.0, 0.0,
               0.2, 0.8), nrow = 2, byrow = TRUE)
classes = matrix(c(1,      2,    3,   4,# j  
                   0.5, 0.25, 0.15, 0.1,# p_j
                   1,    1.9,  2.5, 3), # mu_j
                   nrow = 3, byrow = TRUE)
m = build_model(lambda, N, classes, f, P_a, P_d)
# mean customers in system and mean queue length:
m$mean_clients
m$mean_queue
# Distribution from level 0 to level 20:
dist = m$distribution(20)
}
