% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_ModelTransient}
\alias{Rcpp_ModelTransient}
\alias{ModelTransient}
\alias{Rcpp_ModelTransient-class}
\title{ModelTransient Class}
\description{
Internal representation for multi-server job model with random speed scaling for transient analysis.
}
\section{Slots}{

\describe{
\item{\code{h}}{- Returns current step size.}

\item{\code{mean_clients.}}{\itemize{
\item Parameter: max_time - maximum time for which you want to calculate mean customers in system;
\item Parameter: pi_0 - list of vectors contains distribution at zero time.
\item Returns: mean clients in system.
}}

\item{\code{mean_queue}}{\itemize{
\item Parameter: max_time - maximum time for which you want to calculate mean queue length;
\item Parameter: pi_0 - list of vectors contains distribution at zero time.
\item Returns: mean queue length
}}

\item{\code{distribution}}{\itemize{
\item Parameter: max_time - maximum level for which you want to calculate distribution;
\item Parameter: pi_0 - list of vectors contains distribution at zero time.
\item Returns: list of lists of vectors containing transient distribution.
}}
}}

